requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==3.2
requests==2.26.0
urllib3==1.26.6
python-telegram-bot==13.7
pyTelegramBotAPI==4.0.0
